refactor(dashboard): reuse provider selection menu
