+++ dashboard/src/i18n/locales/en-US/core/shared.json
@@ -34,6 +34,9 @@
   },
   "providerSelector": {
     "notSelected": "Not selected",
+    "defaultModel": "Default model",
+    "model": "Model",
+    "searchPlaceholder": "Search models",
     "buttonText": "Select Provider...",
     "dialogTitle": "Select Provider",
     "loading": "Loading...",
@@ -46,6 +49,7 @@
     "createProvider": "Create Provider",
     "manageProviders": "Provider Management",
     "selectProviderPool": "Select Provider Pool...",
+    "selectedModelCount": "{count} model(s) selected",
     "selectedCount": "{count} provider(s) selected"
   },
   "personaSelector": {
